% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{ctl.hist}
\alias{ctl.hist}
\title{Ritar histogram över överlevnadstiden i år för kontrollgruppen.}
\usage{
ctl.hist()
}
\arguments{
\item{Inga}{Inga inparametrar}
}
\value{
Inga returvärden
}
\description{
Ritar histogram över överlevnadstiden i år för kontrollgruppen.
}
