% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loaddata.R
\name{loadRemote}
\alias{loadRemote}
\title{Ladda data från extern källa.}
\usage{
loadRemote()
}
\value{
Returnerar inläst dataset.
}
\description{
Ladda data från extern källa.
}
\examples{
pbcdata <- loadRemote()
}
