% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{mf.pie}
\alias{mf.pie}
\title{Pie-diagram över fördelningen mellan kvinnor och män i försöksgruppen.}
\usage{
mf.pie()
}
\arguments{
\item{Inga}{Inga inparametrar}
}
\value{
Inga returvärden
}
\description{
Pie-diagram över fördelningen mellan kvinnor och män i försöksgruppen.
}
