% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pie.R
\name{status.pie}
\alias{status.pie}
\title{Pie-diagram över fördelningen mellan transplanterade, avlidna och överlevande (vid studietidens slut)}
\usage{
status.pie()
}
\arguments{
\item{Inga}{Inga inparametrar}
}
\value{
Inga returvärden
}
\description{
Pie-diagram över fördelningen mellan transplanterade, avlidna och överlevande (vid studietidens slut)
}
