% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.R
\name{trt.hist}
\alias{trt.hist}
\title{Ritar histogram över överlevnadstiden i år för den behandlade gruppen.}
\usage{
trt.hist()
}
\arguments{
\item{Inga}{Inga inparametrar}
}
\value{
Inga returvärden
}
\description{
Ritar histogram över överlevnadstiden i år för den behandlade gruppen.
}
