% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trt.R
\name{trt.survival}
\alias{trt.survival}
\title{{ Undersöker om det föreligger en skillnad i överlevnadstid mellan lm-behandlade och placebo-behandlade.}}
\usage{
trt.survival()
}
\arguments{
\item{Inga}{Inga inparametrar}
}
\value{
Inga returvärden
}
\details{
{
Hypotes:
  
  h0 - ingen skillnad föreligger
  
  h1 - det finns en skillnad

  Metod: Mann-Whitney-Wilcoxon-test, då vi inte kan utgå från normalfördelat underlag. 

  Funktionen skriver dels ut resultatet från MWW samt ritar en boxplot med de jämförda subseten.}
}
